tumbuh_sehat_mvp/backend/requirements.txt
# ── Core Framework ─────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# ── Database ───────────────────────────
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# ── Validation & Config ───────────────
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# ── AI / Gemini ────────────────────────
google-generativeai>=0.4.0

# ── Logging ────────────────────────────
python-json-logger>=2.0.0